pyproject.toml
[project]
name = "googleadk-agent-seo-mcp"
version = "0.1.0"
description = "Google ADK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.5.0",
]
